+++ src/components/big-play/big-play.ts
@@ -3,8 +3,8 @@ import svgBigPlay from 'shared/assets/svgs/big-play.svg'
 
 /**
  * Big play template
- * @returns {String} Generated HTML
+ * @returns Generated HTML
  */
 export default function bigPlay(): string {
-	return `<button class="v-bigPlay v-controlButton">${svgBigPlay}</button>`
+	return `<button class="v-bigPlay">${svgBigPlay}</button>`
 }