+++ src/components/poster/poster.ts
@@ -2,9 +2,9 @@ import './poster.css'
 
 /**
  * Poster template
- * @param {Object} options
- * @param {Object} options.posterUrl Poster url
- * @returns {String} Generated HTML
+ * @param options
+ * @param options.posterUrl Poster url
+ * @returns Generated HTML
  */
 export default function poster({ posterUrl = '' }: { posterUrl: string }): string {
 	const posterStyle = posterUrl && ` style="background-image: url(${posterUrl})"`