+++ src/core/player.ts
@@ -1,5 +1,5 @@
 import { formatVideoTime, isTouch } from 'shared/utils/utils.js'
-import { Options, playerParameters, configEvent } from 'shared/assets/interfaces/interfaces.js'
+import { Options, playerParameters, configEvent } from 'shared/assets/types/types.js'
 import ControlBar from 'components/control-bar/control-bar.js'
 
 /**
@@ -16,14 +16,13 @@ export default class Player {
 	isFullScreen: boolean
 	isMuted: boolean
 	isPaused: null | boolean
-	controlBar: any
-	playerEvents: Array<configEvent>
+	controlBar: ControlBar
+	playerEvents: configEvent[]
 	isTouch: boolean
-	plugins: {
-		[key: string]: any
-	}
+	plugins: Record<string, any>
 
 	elements: {
+		outerContainer: HTMLElement
 		container: HTMLElement
 		bigPlay: HTMLElement | null
 		poster: HTMLElement | null
@@ -38,9 +37,9 @@ export default class Player {
 
 	/**
 	 * @constructor
-	 * @param {Object} options
-	 * @param {Class} options.Vlitejs Vlitejs instance
-	 * @param {HTMLElement} options.type Player type (video|audio)
+	 * @param options
+	 * @param options.Vlitejs Vlitejs instance
+	 * @param options.type Player type (video|audio)
 	 */
 	constructor({ Vlitejs, type }: playerParameters) {
 		this.Vlitejs = Vlitejs
@@ -53,6 +52,7 @@ export default class Player {
 
 		this.elements = {
 			container: Vlitejs.container,
+			outerContainer: Vlitejs.outerContainer,
 			bigPlay: Vlitejs.container.querySelector('.v-bigPlay'),
 			poster: Vlitejs.container.querySelector('.v-poster'),
 			controlBar: null,
@@ -187,6 +187,7 @@ export default class Player {
 	 */
 	onReady() {
 		this.options.muted && this.mute()
+		this.media.classList.add('v-media')
 
 		// The iframe needs to be ignored by the focus
 		this.media.setAttribute('tabindex', '-1')
@@ -212,8 +213,8 @@ export default class Player {
 
 	/**
 	 * Add media action listeners on the container
-	 * @param {String} type Event type
-	 * @param {EventListener} listener Event listener
+	 * @param type Event type
+	 * @param listener Event listener
 	 */
 	on(type: string, listener: EventListener) {
 		if (listener instanceof Function) {
@@ -224,8 +225,8 @@ export default class Player {
 
 	/**
 	 * Remove media action listeners on the container
-	 * @param {String} type Event type
-	 * @param {EventListener} listener Event listener
+	 * @param type Event type
+	 * @param listener Event listener
 	 */
 	off(type: string, listener: EventListener) {
 		if (listener instanceof Function) {
@@ -235,10 +236,10 @@ export default class Player {
 
 	/**
 	 * Dispatch custom event on the container
-	 * @param {String} type Event type
-	 * @param {Object} detail Event detail
+	 * @param type Event type
+	 * @param detail Event detail
 	 */
-	dispatchEvent(type: string, detail?: any) {
+	dispatchEvent(type: string, detail?: unknown) {
 		this.elements.container.dispatchEvent(
 			new window.CustomEvent(type, {
 				detail
@@ -248,10 +249,10 @@ export default class Player {
 
 	/**
 	 * Update the loader status
-	 * @param {Boolean} state Status of the loader
+	 * @param state Status of the loader
 	 */
 	loading(state: boolean) {
-		this.elements.container.classList[state ? 'add' : 'remove']('v-loading')
+		this.elements.outerContainer.classList[state ? 'add' : 'remove']('v-loading')
 		this.dispatchEvent('progress')
 	}
 
@@ -271,10 +272,10 @@ export default class Player {
 
 	/**
 	 * Update the progress bar
-	 * @param {Object} options
-	 * @param {String} options.seconds Current time in seconds
-	 * @param {String} options.duration Duration in seconds
-	 * @param {Boolean} options.isRemote Cast mode is enabled
+	 * @param options
+	 * @param options.seconds Current time in seconds
+	 * @param options.duration Duration in seconds
+	 * @param options.isRemote Cast mode is enabled
 	 */
 	updateProgressBar({
 		seconds,
@@ -307,8 +308,8 @@ export default class Player {
 		if (this.options.loop) {
 			this.play()
 		} else {
-			this.elements.container.classList.replace('v-playing', 'v-paused')
-			this.elements.container.classList.add('v-firstStart')
+			this.elements.outerContainer.classList.replace('v-playing', 'v-paused')
+			this.elements.outerContainer.classList.add('v-firstStart')
 		}
 
 		if (this.elements.poster) {
@@ -335,7 +336,7 @@ export default class Player {
 		if (this.isLinearAd) return
 
 		if (this.isPaused === null) {
-			this.elements.container.classList.remove('v-firstStart')
+			this.elements.outerContainer.classList.remove('v-firstStart')
 
 			if (this.type === 'video' && this.elements.poster) {
 				this.elements.poster.classList.remove('v-active')
@@ -344,7 +345,7 @@ export default class Player {
 
 		!this.isCast && this.methodPlay()
 		this.isPaused = false
-		this.elements.container.classList.replace('v-paused', 'v-playing')
+		this.elements.outerContainer.classList.replace('v-paused', 'v-playing')
 
 		if (this.elements.playPause) {
 			this.elements.playPause.setAttribute('aria-label', 'Pause')
@@ -365,7 +366,7 @@ export default class Player {
 	pause() {
 		!this.isCast && this.methodPause()
 		this.isPaused = true
-		this.elements.container.classList.replace('v-playing', 'v-paused')
+		this.elements.outerContainer.classList.replace('v-playing', 'v-paused')
 
 		if (this.elements.playPause) {
 			this.elements.playPause.setAttribute('aria-label', 'Play')
@@ -392,7 +393,7 @@ export default class Player {
 
 	/**
 	 * Set player volume
-	 * @param {Number} volume New volume
+	 * @param volume New volume
 	 */
 	setVolume(volume: number) {
 		if (volume > 1) {
@@ -418,7 +419,7 @@ export default class Player {
 
 	/**
 	 * Get player volume
-	 * @returns {Promise<Number>} Player volume
+	 * @returns Player volume
 	 */
 	getVolume(): Promise<number> {
 		return new window.Promise((resolve) => {
@@ -460,7 +461,7 @@ export default class Player {
 
 	/**
 	 * Update the current time of the media element
-	 * @param {Number} newTime New current time of the media element
+	 * @param newTime New current time of the media element
 	 */
 	seekTo(newTime: number) {
 		this.methodSeekTo(newTime)
@@ -478,7 +479,7 @@ export default class Player {
 			// @ts-ignore: Object is possibly 'null'.
 			this.elements.container[requestFn]()
 			this.isFullScreen = true
-			this.elements.container.classList.add('v-fullscreenButtonDisplay')
+			this.elements.outerContainer.classList.add('v-fullscreenButtonDisplay')
 
 			if (this.elements.fullscreen) {
 				this.elements.fullscreen.classList.add('v-controlPressed')
@@ -491,8 +492,8 @@ export default class Player {
 
 	/**
 	 * Exit the fullscreen
-	 * @param {Object} options
-	 * @param {Boolean} options.escKey The exit is trigger by the esk key
+	 * @param options
+	 * @param options.escKey The exit is trigger by the esk key
 	 */
 	exitFullscreen({ escKey = false }: { escKey?: boolean } = {}) {
 		const { cancelFn } = this.Vlitejs.supportFullScreen
@@ -502,7 +503,7 @@ export default class Player {
 			!escKey && document[cancelFn]()
 			this.isFullScreen = false
 
-			this.elements.container.classList.remove('v-fullscreenButtonDisplay')
+			this.elements.outerContainer.classList.remove('v-fullscreenButtonDisplay')
 
 			if (this.elements.fullscreen) {
 				this.elements.fullscreen.classList.remove('v-controlPressed')